Remove githubId param from GET /api/dashboards/:githubId modified:   src/client/app/home/home.js modified:   src/client/app/services/services.js modified:   src/server/config/routes.js deleted:    src/server/request-handlers/dashboards-githubId.js modified:   src/server/request-handlers/dashboards.js
